avoid-duplicate-files initial commit
Added -A and -s to avoid redownloading the same file that has a similar filename. Files that share the same track name will be detected. If -A is used, it will ask if the user wants to skip the file or not. If -s is used, it will skip the file without asking for input.